Breaking down the CupertinoSwipeBox components into separate files
